SLING-11229 adding shared test classes
