+++ src/test/java/org/apache/sling/jcr/resource/internal/helper/jcr/JcrResourceProviderTest.java
@@ -18,7 +18,6 @@
  */
 package org.apache.sling.jcr.resource.internal.helper.jcr;
 
-import java.lang.annotation.Annotation;
 import java.security.Principal;
 
 import javax.jcr.Node;
@@ -29,7 +28,6 @@ import javax.jcr.nodetype.NodeType;
 
 import org.apache.sling.api.resource.PersistenceException;
 import org.apache.sling.api.resource.Resource;
-import org.apache.sling.jcr.resource.internal.helper.jcr.JcrResourceProvider.Config;
 import org.apache.sling.spi.resource.provider.ResolveContext;
 import org.apache.sling.spi.resource.provider.ResourceContext;
 import org.junit.Assert;
@@ -51,20 +49,7 @@ public class JcrResourceProviderTest extends SlingRepositoryTestBase {
         ComponentContext ctx = Mockito.mock(ComponentContext.class);
         Mockito.when(ctx.locateService(Mockito.anyString(), Mockito.any(ServiceReference.class))).thenReturn(repo);
         jcrResourceProvider = new JcrResourceProvider();
-        jcrResourceProvider.activate(ctx, new Config() {
-            @Override
-            public Class<? extends Annotation> annotationType() {
-                return null;
-            }
-            @Override
-            public boolean enable_query_limit() {
-                return false;
-            }
-            @Override
-            public long query_limit() {
-                return 0;
-            }
-        });
+        jcrResourceProvider.activate(ctx, new SimpleConfig(false, -1));
     }
 
     @Override
