+++ src/test/java/org/apache/sling/jcr/resource/internal/helper/jcr/SimpleConfig.java
@@ -1,0 +1,47 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.sling.jcr.resource.internal.helper.jcr;
+
+import java.lang.annotation.Annotation;
+
+public class SimpleConfig implements JcrResourceProvider.Config {
+    private final boolean enabled;
+    private final long limit;
+
+    public SimpleConfig(boolean enabled, long limit) {
+        this.enabled = enabled;
+        this.limit = limit;
+    }
+
+    @Override
+    public Class<? extends Annotation> annotationType() {
+        return null;
+    }
+
+    @Override
+    public boolean enable_query_limit() {
+        return enabled;
+    }
+
+    @Override
+    public long query_limit() {
+        return this.limit;
+    }
+
+}
